requirements.txt
openai>=1.0.0
python-dotenv==1.0.0
pandas>=1.5.0
colorlog==6.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
numpy>=1.23.0
tqdm>=4.64.0 